refactor(wave4-step2x): thin cache facade for read/evict/list paths
