+++ crates/assay-registry/src/cache_next/mod.rs
@@ -6,9 +6,11 @@
 //! - No behavior/perf changes in this commit.
 
 pub(crate) mod errors;
+pub(crate) mod evict;
 pub(crate) mod integrity;
 pub(crate) mod io;
 pub(crate) mod keys;
 pub(crate) mod policy;
 pub(crate) mod put;
+pub(crate) mod read;
 pub(crate) mod tests;
