+++ docs/architecture/PLAN-split-refactor-2026q1.md
@@ -299,7 +299,8 @@ Security improvements:
 Step status:
 
 - Step 1 (`lockfile.rs` + `cache.rs` behavior freeze/inventory/gates): merged via PR #339.
-- Step 2 (`lockfile.rs` + `cache.rs` mechanical split): in progress on `codex/wave4-step2-lockfile-cache-split`.
+- Step 2 (`lockfile.rs` + `cache.rs` mechanical split): merged via PR #340.
+- Step 2.x (`cache.rs` facade thinness follow-up for read/evict/list/get_metadata): in progress on `codex/wave4-step2x-cache-thinness`.
 - `explain.rs` follows as a separate slice to keep review size and blast radius small.
 
 ### Objectives
